+++ packages/framework/src/utils/index.ts
@@ -1,5 +1,8 @@
 export { assertUnreachable } from "./assert";
+export { toSnakeCase } from "./case";
 export { readPositiveIntEnv } from "./env-parse";
 export { generateId } from "./ids";
+export { isPlainObject } from "./is-plain-object";
+export { parseStringArrayJson } from "./parse-string-array-json";
 export { parseJsonOrThrow, parseJsonSafe, stringifyJson } from "./safe-json";
 export { parseRoles } from "./serialization";