+++ packages/dispatcher-live/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cosmicdrift/kumiko-dispatcher-live
 
+## 0.16.0
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-headless@0.16.0
+
 ## 0.15.0
 
 ### Patch Changes